prophet-predictor/requirements.txt
flask==2.3.3
prophet==1.1.4
requests==2.31.0
pandas==2.0.3
redis==4.6.0
scikit-learn==1.3.0
numpy==1.24.3
pyyaml==6.0.1
prometheus-client==0.17.1
urllib3==2.0.4
